requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
paddlepaddle-gpu==2.6.0
paddleocr>=2.7.0
opencv-python-headless>=4.5.0
Pillow>=9.0.0
numpy>=1.21.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
